console: Refactor webhook add form
Move from class components to functional components
