+++ pkg/webui/locales/en.json
@@ -568,9 +568,9 @@
   "console.views.application-integrations-pubsub-edit.application-integrations-pubsub-edit.deleteSuccess": "Pub/Sub deleted",
   "console.views.application-integrations-webhook-add-choose.application-integrations-webhook-add-choose.chooseTemplate": "Choose webhook template",
   "console.views.application-integrations-webhook-add-choose.application-integrations-webhook-add-choose.customTileDescription": "Create a custom webhook without template",
-  "console.views.application-integrations-webhook-add-form.index.addCustomWebhook": "Add custom webhook",
-  "console.views.application-integrations-webhook-add-form.index.addWebhookViaTemplate": "Add {templateName} webhook",
-  "console.views.application-integrations-webhook-add-form.index.customWebhook": "Custom webhook",
+  "console.views.application-integrations-webhook-add-form.application-integrations-webhook-add-form.addCustomWebhook": "Add custom webhook",
+  "console.views.application-integrations-webhook-add-form.application-integrations-webhook-add-form.addWebhookViaTemplate": "Add {templateName} webhook",
+  "console.views.application-integrations-webhook-add-form.application-integrations-webhook-add-form.customWebhook": "Custom webhook",
   "console.views.application-integrations-webhook-edit.application-integrations-webhook-edit.editWebhook": "Edit webhook",
   "console.views.application-integrations-webhook-edit.application-integrations-webhook-edit.updateSuccess": "Webhook updated",
   "console.views.application-integrations-webhook-edit.application-integrations-webhook-edit.deleteSuccess": "Webhook deleted",
