+++ pkg/webui/locales/ja.json
@@ -362,9 +362,9 @@
   "console.views.application-integrations-pubsub-edit.application-integrations-pubsub-edit.deleteSuccess": "Pub/Subの削除",
   "console.views.application-integrations-webhook-add-choose.application-integrations-webhook-add-choose.chooseTemplate": "webhookテンプレートの選択",
   "console.views.application-integrations-webhook-add-choose.application-integrations-webhook-add-choose.customTileDescription": "テンプレート無しでカスタムwebhookを作成",
-  "console.views.application-integrations-webhook-add-form.index.addCustomWebhook": "カスタム webhookを追加",
-  "console.views.application-integrations-webhook-add-form.index.addWebhookViaTemplate": "{templateName} webhookを追加します。",
-  "console.views.application-integrations-webhook-add-form.index.customWebhook": "カスタム webhook",
+  "console.views.application-integrations-webhook-add-form.application-integrations-webhook-add-form.addCustomWebhook": "カスタム webhookを追加",
+  "console.views.application-integrations-webhook-add-form.application-integrations-webhook-add-form.addWebhookViaTemplate": "{templateName} webhookを追加します。",
+  "console.views.application-integrations-webhook-add-form.application-integrations-webhook-add-form.customWebhook": "カスタム webhook",
   "console.views.application-integrations-webhook-edit.application-integrations-webhook-edit.editWebhook": "webhookを編集",
   "console.views.application-integrations-webhook-edit.application-integrations-webhook-edit.updateSuccess": "webhookを更新",
   "console.views.application-integrations-webhook-edit.application-integrations-webhook-edit.deleteSuccess": "webhookを削除",
